sql-connect/Snickr Project/requirements.txt
flask
python-dotenv
psycopg2-binary 
werkzeug